moved caption to molecules
