+++ src/atoms/Image/Image.test.tsx
@@ -4,24 +4,10 @@ import { mountWithTheme } from '../../test';
 import Image from './index';
 
 describe('<Image />', () => {
-  describe('Default', () => {
-    it('has the correct style attributes', () => {
-      const img = mountWithTheme(<Image src="img.png" />);
-      expect(img).toHaveStyleRule('border-radius', '2px');
-      expect(img).toHaveStyleRule('width', '100%');
-    });
-  });
+  it('has the correct style attributes', () => {
+    const img = mountWithTheme(<Image src="img.png" />);
+    expect(img).toHaveStyleRule('border-radius', '2px');
 
-  describe('With Caption', () => {
-    it('has the correct style attributes', () => {
-      const img = mountWithTheme(<Image src="img.png" caption="A caption" />);
-      expect(img).toHaveStyleRule('border-radius', '2px');
-      expect(img).toHaveStyleRule('width', '100%');
-    });
-
-    it('displays the caption', () => {
-      const img = mountWithTheme(<Image src="img.png" caption="A caption" />);
-      expect(img.find('div').text()).toBe('A caption');
-    });
+    expect(img).toHaveStyleRule('width', '100%');
   });
 });
