+++ src/molecules/ImageWithCaption/ImageWithCaption.test.tsx
@@ -1,0 +1,28 @@
+import React from 'react';
+import { mountWithTheme } from '../../test';
+import { Caption } from '../../atoms/Typography';
+
+import Image from './index';
+
+describe('<Image />', () => {
+  describe('Default', () => {
+    it('has the correct style attributes', () => {
+      const img = mountWithTheme(<Image src="img.png" />);
+      expect(img).toHaveStyleRule('border-radius', '2px');
+      expect(img).toHaveStyleRule('width', '100%');
+    });
+  });
+
+  describe('With Caption', () => {
+    it('has the correct style attributes', () => {
+      const img = mountWithTheme(<Image src="img.png" caption="A caption" />);
+      expect(img).toHaveStyleRule('border-radius', '2px');
+      expect(img).toHaveStyleRule('width', '100%');
+    });
+
+    it('displays the caption', () => {
+      const img = mountWithTheme(<Image src="img.png" caption="A caption" />);
+      expect(img.find(Caption).text()).toBe('A caption');
+    });
+  });
+});
