+++ src/molecules/ImageWithCaption/README.md
@@ -1,0 +1,20 @@
+# Image With Caption
+
+To implement an `ImageWithCaption` into your project you'll need to add this import
+```js
+import { Image } from '@sofarsounds/maestro'
+```
+
+After adding the import you can use the image simply like this
+```html
+<Image src="image.png" caption="A caption" />
+```
+
+## Props
+Table below contains all types of props available in the Image component  
+**Note:** all other standard image html props apply
+
+| Name          | Type     | Default         | Description                      |
+| :------------ | :-----   | :-------------- | :------------------------------- |
+| **src**       | `string` |                 | Src for the image
+| **caption**   | `string` |                 | Caption to display below the image
